DOMA-1183 extracted `TicketAssignments` into separate module
