+++ apps/condo/domains/ticket/components/BaseTicketForm/index.tsx
@@ -2,17 +2,13 @@
 // @ts-nocheck
 import { useIntl } from '@core/next/intl'
 import { Checkbox, Col, Form, Input, Row, Typography, Tooltip, Tabs } from 'antd'
-import { get, find } from 'lodash'
+import { get } from 'lodash'
 import React, { useEffect, useRef, useState } from 'react'
 import { ITicketFormState } from '@condo/domains/ticket/utils/clientSchema/Ticket'
 import { FormWithAction } from '@condo/domains/common/components/containers/FormList'
 import { PropertyAddressSearchInput } from '@condo/domains/property/components/PropertyAddressSearchInput'
-import { searchEmployeeUser } from '@condo/domains/ticket/utils/clientSchema/search'
 import { FocusContainer } from '@condo/domains/common/components/FocusContainer'
-import { GraphQlSearchInput } from '@condo/domains/common/components/GraphQlSearchInput'
-import { LabelWithInfo } from '@condo/domains/common/components/LabelWithInfo'
 import { UnitNameInput } from '@condo/domains/user/components/UnitNameInput'
-import { UserNameField } from '@condo/domains/user/components/UserNameField'
 import { useTicketValidations } from './useTicketValidations'
 import { FrontLayerContainer } from '@condo/domains/common/components/FrontLayerContainer'
 import { useMultipleFileUploadHook } from '@condo/domains/common/components/MultipleFileUpload'
@@ -25,7 +21,7 @@ import { InputWithCounter } from '@condo/domains/common/components/InputWithCoun
 import Prompt from '@condo/domains/common/components/Prompt'
 import { IOrganizationEmployeeRoleUIState } from '@condo/domains/organization/utils/clientSchema/OrganizationEmployeeRole'
 import { IOrganizationUIState } from '@condo/domains/organization/utils/clientSchema/Organization'
-import { AutoAssignerByDivisions } from './AutoAssignerByDivisions'
+import { TicketAssignments } from './TicketAssignments'
 
 const { TabPane } = Tabs
 
@@ -205,89 +201,6 @@ export const TicketInfo = ({ form, validations, UploadComponent, initialValues, 
                 </Row>
             </Col>
         </>
-    )
-}
-
-const TicketAssignments = ({ validations, organizationId, propertyId, disableUserInteraction, autoAssign, categoryClassifier, form }) => {
-    const intl = useIntl()
-    const TicketAssignmentTitle = intl.formatMessage({ id: 'TicketAssignment' })
-    const ExecutorLabel = intl.formatMessage({ id: 'field.Executor' })
-    const ResponsibleLabel = intl.formatMessage({ id: 'field.Responsible' })
-    const ExecutorExtra = intl.formatMessage({ id: 'field.Executor.description' })
-    const ResponsibleExtra = intl.formatMessage({ id: 'field.Responsible.description' })
-
-    const formatUserFieldLabel = ({ text, value, data: employee }) => {
-        const matchedSpecialization = find(employee.specializations, { id: categoryClassifier })
-        return (
-            <UserNameField user={{ name: text, id: value }}>
-                {({ name, postfix }) => (
-                    <>
-                        <Typography.Text>
-                            {name} {postfix}
-                        </Typography.Text>
-                        {matchedSpecialization && (
-                            <Typography.Text type="secondary">
-                                {`(${matchedSpecialization.name.toLowerCase()})`}
-                            </Typography.Text>
-                        )}
-                    </>
-                )}
-            </UserNameField>
-        )
-    }
-
-    return (
-        <Col span={24}>
-            <Row justify={'space-between'} gutter={[0, 24]}>
-                <Col span={24}>
-                    <Typography.Title level={5} style={{ margin: '0' }}>{TicketAssignmentTitle}</Typography.Title>
-                </Col>
-                {autoAssign && propertyId && (
-                    <Col span={24}>
-                        <AutoAssignerByDivisions
-                            organizationId={organizationId}
-                            propertyId={propertyId}
-                            categoryClassifier={categoryClassifier}
-                            form={form}
-                        />
-                    </Col>
-                )}
-                <Col span={11}>
-                    <Form.Item
-                        name={'executor'}
-                        rules={validations.executor}
-                        label={<LabelWithInfo title={ExecutorExtra} message={ExecutorLabel}/>}
-                    >
-                        <GraphQlSearchInput
-                            formatLabel={formatUserFieldLabel}
-                            search={searchEmployeeUser(organizationId, ({ role }) => (
-                                get(role, 'canBeAssignedAsExecutor', false)
-                            ))}
-                            allowClear={false}
-                            showArrow={false}
-                            disabled={disableUserInteraction}
-                        />
-                    </Form.Item>
-                </Col>
-                <Col span={11}>
-                    <Form.Item
-                        name={'assignee'}
-                        rules={validations.assignee}
-                        label={<LabelWithInfo title={ResponsibleExtra} message={ResponsibleLabel}/>}
-                    >
-                        <GraphQlSearchInput
-                            formatLabel={formatUserFieldLabel}
-                            search={searchEmployeeUser(organizationId, ({ role }) => (
-                                get(role, 'canBeAssignedAsResponsible', false)
-                            ))}
-                            allowClear={false}
-                            showArrow={false}
-                            disabled={disableUserInteraction}
-                        />
-                    </Form.Item>
-                </Col>
-            </Row>
-        </Col>
     )
 }
